src/requirements.txt
keras
minio
numpy
pandas
python-dotenv
scikit_learn
tensorflow
yfinance
